+++ CHANGELOG.md
@@ -69,6 +69,7 @@ This is a work-in-progress.
 * QuPath doesn't support some channel combinations through Bio-Formats (https://github.com/qupath/qupath/issues/956)
 * 'Cannot invoke “java.lang.Double.doubleValue()”' exception in 'Create Thresholder' (https://github.com/qupath/qupath/issues/988)
 * Uncaught exceptions can fill the screen with duplicate error notifications (https://github.com/qupath/qupath/issues/990)
+* Locked point annotations can still be edited (https://github.com/qupath/qupath/issues/1001)
 
 ### Changes through Bio-Formats 6.10.0
 * Bio-Formats 6.10.0 brings several important new features to QuPath, including: